Objects - Game data refactored - issue with character login (#355)
* wip

* WIP - character login is trying to persist a world object

* Still throwing an exception on player login.

* Working Code - finished GameData - need to complete PhysicsData and ModelData.   Thanks to @Lidefeath for finding my issue.

* WIP compiles but has error on run.   Data changes have broken something.   Will fix tomorrow.

* Wip - nothing really working yet.   I can log in but item creation is not working.

+++ Source/ACE.Entity/AceObjectPropertiesDouble.cs
@@ -10,7 +10,7 @@ namespace ACE.Entity
         public uint AceObjectId { get; set; }
 
         [DbField("dblPropertyId", (int)MySqlDbType.UInt16)]
-        public uint DblPropertyId { get; set; }
+        public short DblPropertyId { get; set; }
 
         [DbField("propertyValue", (int)MySqlDbType.Double)]
         public double PropertyValue { get; set; }
